requirements.txt
fastapi
uvicorn[standard]
python-can
python-can-csscan-serial
canalystii
cantools
pyserial
pydantic
pytest
httpx
